code refactoring, moved model classes to the model folder
