+++ proactive/ProactiveHelper.py
@@ -1,9 +1,9 @@
-from .ProactiveScriptLanguage import *
 from .ProactiveFactory import *
-from .ProactiveTask import *
-from .ProactiveJob import *
 from .ProactiveBuilder import *
 
+from .model.ProactiveScriptLanguage import *
+from .model.ProactiveTask import *
+from .model.ProactiveJob import *
 
 class ProactiveHelper:
   """
