+++ proactive/__init__.py
@@ -1,7 +1,8 @@
 from .ProactiveGateway import *
 from .ProactiveHelper import *
-from .ProactiveScriptLanguage import *
 from .ProactiveFactory import *
-from .ProactiveTask import *
-from .ProactiveJob import *
 from .ProactiveBuilder import *
+
+from .model.ProactiveScriptLanguage import *
+from .model.ProactiveTask import *
+from .model.ProactiveJob import *
